Add Safari iOS support
Currently there are too many serious bugs affecting Safari on iOS for it to be used in production by anyone. Therefore we must add a not supported for Safari on iOS specifically. Attempt to use it at your own risk!

Bugs that are currently breaking production use of Safari on iOS:
- https://bugs.webkit.org/show_bug.cgi?id=179964
- https://bugs.webkit.org/show_bug.cgi?id=181009
- https://bugs.webkit.org/show_bug.cgi?id=180748

List of bugs for Safari WebRTC:
https://bugs.webkit.org/buglist.cgi?bug_status=__open__&component=WebRTC&list_id=3330069&product=WebKit